multi_agent_debate_starter/requirements.txt
openai>=1.40.0
pydub>=0.25.1
pyyaml>=6.0.1
